requirements.txt
gitpython>=3.1.40
openai>=1.12.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
click>=8.1.0
pytest>=8.0.0
ruff>=0.4.0
